+++ packages/api-core/src/exports/api/buildParam.ts
@@ -1,1 +1,1 @@
-export { BuildParam, BuildParamRegistry } from "~/features/buildParam/index.js";
+export { BuildParam, BuildParams } from "~/features/buildParams/index.js";
