--- packages/api-core/src/features/buildParam/index.ts
@@ -1,1 +1,0 @@
-export { BuildParam, BuildParamRegistry } from "./abstractions.js";
