+++ packages/project/src/defineExtension/index.ts
@@ -1,5 +1,4 @@
 export * from "./defineExtension.js";
 export * from "./defineApiExtension.js";
-export * from "./defineApiBuildParamExtension.js";
 export * from "./models/index.js";
 export * from "./zodTypes/zodSrcPath.js";
